requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.20.0
alembic==1.13.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
billiard==4.2.4
blis==1.3.3
build==1.4.0
catalogue==2.0.10
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
cmdstanpy==1.3.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
datasets==4.8.2
dill==0.4.1
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.109.2
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.1.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
holidays==0.93
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kombu==5.6.2
kubernetes==35.0.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.3
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.4
opencv-python==4.11.0.86
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
polars==1.39.3
polars-runtime-32==1.39.3
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.1.5
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
py-vapid==1.9.4
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2026.1.post1
pywebpush==2.0.0
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.17.1
seaborn==0.13.2
sentence-transformers==5.3.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.5.2
stanio==0.5.1
starlette==0.36.3
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.3
transformers==4.57.6
twilio==8.13.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.1.29
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.22.1
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
